+++ pyjobshop/__init__.py
@@ -1,6 +1,13 @@
 from .constants import MAX_VALUE as MAX_VALUE
 from .Model import Model as Model
+from .ProblemData import Consecutive as Consecutive
 from .ProblemData import Constraints as Constraints
+from .ProblemData import DifferentResources as DifferentResources
+from .ProblemData import EndAtEnd as EndAtEnd
+from .ProblemData import EndAtStart as EndAtStart
+from .ProblemData import EndBeforeEnd as EndBeforeEnd
+from .ProblemData import EndBeforeStart as EndBeforeStart
+from .ProblemData import IdenticalResources as IdenticalResources
 from .ProblemData import Job as Job
 from .ProblemData import Machine as Machine
 from .ProblemData import Mode as Mode
@@ -8,6 +15,10 @@ from .ProblemData import NonRenewable as NonRenewable
 from .ProblemData import Objective as Objective
 from .ProblemData import ProblemData as ProblemData
 from .ProblemData import Renewable as Renewable
+from .ProblemData import StartAtEnd as StartAtEnd
+from .ProblemData import StartAtStart as StartAtStart
+from .ProblemData import StartBeforeEnd as StartBeforeEnd
+from .ProblemData import StartBeforeStart as StartBeforeStart
 from .ProblemData import Task as Task
 from .read import read as read
 from .Result import Result as Result
